+++ src/AirdropServiceProvider.php
@@ -15,7 +15,7 @@ use Illuminate\Support\ServiceProvider;
 
 class AirdropServiceProvider extends ServiceProvider
 {
-    public function boot()
+    public function boot(): void
     {
         if ($this->app->runningInConsole()) {
             $this->commands([
@@ -32,7 +32,7 @@ class AirdropServiceProvider extends ServiceProvider
         ], 'config');
     }
 
-    public function register()
+    public function register(): void
     {
         $this->mergeConfigFrom(__DIR__ . '/../config/airdrop.php', 'airdrop');
     }