+++ src/Commands/Install.php
@@ -12,18 +12,8 @@ use Illuminate\Support\Facades\Artisan;
 
 class Install extends Command
 {
-    /**
-     * The name and signature of the console command.
-     *
-     * @var string
-     */
     protected $signature = 'airdrop:install';
 
-    /**
-     * The console command description.
-     *
-     * @var string
-     */
     protected $description = 'Install the Airdrop config file into your app.';
 
     public function __construct()
@@ -35,10 +25,7 @@ class Install extends Command
         }
     }
 
-    /**
-     * @throws Exception
-     */
-    public function handle()
+    public function handle(): void
     {
         Artisan::call('vendor:publish', [
             '--provider' => AirdropServiceProvider::class