+++ src/Contracts/TriggerContract.php
@@ -11,9 +11,6 @@ interface TriggerContract
     /**
      * Return any state that should be considered when determining
      * whether or not your build process needs to run again.
-     *
-     * @param  array  $config
-     * @return array
      */
-    public function triggerBuildWhenChanged($config = []);
+    public function triggerBuildWhenChanged(array $config = []): array;
 }